Add Xcode project.
FossilOrigin-Name: cf2fe185976eb9af8c8437fb68da475cdf47944bef02a89114dc3ea285b97552

+++ PGConnection.m
@@ -84,7 +84,7 @@
 	PGresult *result;
 	const char **values;
 	va_list args, args2;
-	size_t argsCount;
+	int argsCount;
 
 	va_start(args, parameter);
 	va_copy(args2, args);
